16.09.2024

+++ 08-API/02-fetch/js/script.js
@@ -1,9 +1,60 @@
-import { deleteUser, getAllUsers } from "./user-api.js";
-
+import {
+	createUser,
+	deleteUser,
+	getAllUsers,
+	getUserById,
+    updateUser,
+} from "./user-api.js";
 
 const tblUsers = document.getElementById("tblUsers");
 const tbodyUsers = tblUsers.querySelector("tbody");
+const frmUser = document.getElementById("frmUser");
+const txtFirstName = document.getElementById("txtFirstName");
+const txtLastName = document.getElementById("txtLastName");
+const txtEmail = document.getElementById("txtEmail");
+const btnSubmit = document.getElementById("btnSubmit");
 
+const addUser = async () => {
+	const firstName = txtFirstName.value;
+	const lastName = txtLastName.value;
+	const email = txtEmail.value;
+
+	const user = {
+		firstName,
+		lastName,
+		email,
+	};
+
+	await createUser(user);
+	init();
+	frmUser.reset();
+};
+
+const editUser = async (id) => {
+	const user = await getUserById(id);
+	const { firstName, lastName, email, avatar } = user;
+	txtFirstName.value = firstName;
+	txtLastName.value = lastName;
+	txtEmail.value = email;
+};
+
+const saveUser = async (id) =>{
+    const firstName = txtFirstName.value;
+	const lastName = txtLastName.value;
+	const email = txtEmail.value;
+
+	const user = {
+		firstName,
+		lastName,
+		email,
+	};
+
+	await updateUser(id, user);
+	init();
+	frmUser.reset();
+    frmUser.dataset.method = "create";
+    btnSubmit.innerHTML = "➕ Add"
+}
 
 const renderUserList = (users) => {
 	let strUsers = "";
@@ -11,33 +62,57 @@ const renderUserList = (users) => {
 	users.forEach((item, index) => {
 		strUsers += `<tr>
                         <td>${index + 1}</td>
-                        <td><img src="${item.avatar}" alt="${item.firstName}"></td>
+						<td><img src="${item.avatar}"></td>
                         <td>${item.firstName}</td>
                         <td>${item.lastName}</td>
                         <td>${item.email}</td>
-                        <td><button class="btn btn-danger btn-sm btn-del" data-id="${item.id}">🗑️</button></td>
+                        <td style="width:100px; text-nowrap">
+                            <button class="btn btn-info btn-sm btn-edit" data-id="${
+								item.id
+							}">✍️</button>
+                            <button class="btn btn-danger btn-sm btn-del" data-id="${
+								item.id
+							}">🗑️</button></td>
                     </tr>`;
 	});
 	return strUsers;
 };
 
-
-
-const init = async () => { 
-    const users = await getAllUsers();  
-    const strUsers = renderUserList(users);
-    tbodyUsers.innerHTML = strUsers;
-}
-
+const init = async () => {
+	const users = await getAllUsers();
+	const strUsers = renderUserList(users);
+	tbodyUsers.innerHTML = strUsers;
+};
 
 init();
 
+tbodyUsers.addEventListener("click", async (e) => {
+	const userId = e.target.dataset.id;
+	if (!userId) return;
 
-tbodyUsers.addEventListener("click", async (e)=>{
+	if (e.target.classList.contains("btn-del")) { // olayin oldugu yerin classlistinde btn-del varsa
+		const res = confirm("Are you sure to delete?");
+		if (!res) return;
+		await deleteUser(userId); // deleteUser async oldugu icin awaitle bekletmemiz lazim. 
+		init();
+	}
+	else if (e.target.classList.contains("btn-edit")) {
+		editUser(userId);
+		window.scrollTo(0, 0);
+		btnSubmit.innerHTML = "💾 Update";
+		frmUser.dataset.method = "update";
+		frmUser.dataset.id = userId;
+	}
+});
 
-    if(e.target.classList.contains("btn-del")){
-        const userId = e.target.dataset.id;
-        await deleteUser(userId)
-        init()
-    }
-})
+frmUser.addEventListener("submit", (e) => {
+	e.preventDefault(); // submit butonunun default davranisini onle diyoruz
+	const method = e.target.dataset.method;
+	const userId = e.target.dataset.id;
+
+	if (method === "create") {
+		addUser();
+	} else {
+        saveUser(userId)
+	}
+});
